+++ src/oidc-agent/oidcd/parse_internal.c
@@ -1,6 +1,7 @@
 #include "parse_internal.h"
 
 #include "defines/ipc_values.h"
+#include "utils/agentLogger.h"
 #include "utils/json.h"
 #include "utils/memory.h"
 #include "utils/oidc_error.h"
@@ -62,4 +63,31 @@ oidc_error_t parseForErrorCode(char* res) {
     return oidc_errno;
   }
   return OIDC_SUCCESS;
+}
+
+char* parseStateLookupRes(char* res) {
+  INIT_KEY_VALUE(IPC_KEY_STATUS, IPC_KEY_CONFIG, OIDC_KEY_ERROR);
+  if (CALL_GETJSONVALUES(res) < 0) {
+    secFree(res);
+    return NULL;
+  }
+  KEY_VALUE_VARS(status, config, error);
+  secFree(res);
+  if (_error != NULL) {
+    agent_log(ERROR, _error);
+    SEC_FREE_KEY_VALUES();
+    return NULL;
+  }
+  if (_config) {
+    char* config = oidc_strcopy(_config);
+    SEC_FREE_KEY_VALUES();
+    return config;
+  }
+  if (strcaseequal(_status, STATUS_NOTFOUND)) {
+    SEC_FREE_KEY_VALUES();
+    oidc_errno = OIDC_EWRONGSTATE;
+    return NULL;
+  }
+  SEC_FREE_KEY_VALUES();
+  return NULL;
 }