+++ src/oidc-gen/gen_handler.h
@@ -25,7 +25,7 @@ void  stateLookUpWithConfigSave(const char*             state,
 char* configFromStateLookUp(const char*             state,
                             const struct arguments* arguments);
 void  gen_handlePrint(const char* file, const struct arguments* arguments);
-char* gen_handleDeviceFlow(char* json_device, char* json_account,
+char* gen_handleDeviceFlow(const char*             json_device,
                            const struct arguments* arguments);
 oidc_error_t gen_handlePublicClient(struct oidc_account* account,
                                     struct arguments*    arguments);