+++ src/oidc-gen/parse_ipc.h
@@ -2,6 +2,6 @@
 #define PARSE_IPC_GEN_H
 
 #include "oidc-gen/oidc-gen_options.h"
-char* gen_parseResponse(char* res, const struct arguments* arguments);
+char* gen_parseResponse(char* res, const struct arguments* argumentsg);
 
 #endif  // PARSE_IPC_GEN_H