+++ src/oidc-gen/qr.h
@@ -4,5 +4,6 @@
 char* getQRCode(const char* content);
 char* getUTF8QRCode(const char* content);
 char* getANSIQRCode(const char* content);
+int   getIMGQRCode(const char* content, const char* outpath);
 
 #endif  // OIDC_AGENT_QR_H