+++ src/utils/oidc_error.c
@@ -104,6 +104,7 @@ char* oidc_serrorFor(oidc_error_t err) {
     case OIDC_ENOCODE: return "redirected uri did not contain code parameter";
     case OIDC_ENOBASEURI: return "could not get base uri from redirected uri";
     case OIDC_EWRONGSTATE: return "wrong state";
+    case OIDC_EWRONGDEVICECODE: return "wrong device code";
     case OIDC_ENOPRIVCONF: return "Privilege configuration file not found";
     case OIDC_ENOSUPREG:
       return "Dynamic registration is not supported by this issuer. Please "