+++ src/utils/system_runner.c
@@ -3,6 +3,7 @@
 
 #include <stdio.h>
 #include <stdlib.h>
+#include <unistd.h>
 
 #include "utils/file_io/file_io.h"
 #include "utils/logger.h"
@@ -25,4 +26,19 @@ char* getOutputFromCommand(const char* cmd) {
   char* ret = readFILE(fp);
   pclose(fp);
   return ret;
+}
+
+void fireCommand(const char* cmd) {
+  pid_t pid = fork();
+  if (pid == -1) {
+    logger(ERROR, "fork %m");
+    return;
+  } else if (pid > 0) {  // parent
+    return;
+  }
+  // child
+  execlp("/bin/sh", "sh", "-c", cmd, (char*)NULL);
+  /* exec functions only return on error */
+  logger(ERROR, "Error executing command: %m");
+  exit(EXIT_FAILURE);
 }